requirements.txt
yfinance
panel
hvplot
bokeh
requests
python-dotenv

